+++ protected/humhub/docs/RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 1.5
 ------------------------------------
 
+- Admin log filter (https://github.com/humhub/humhub/issues/3909) (TBD)
 - SameSite Cookie Support (TBD)